support revenge after network game (fixes #55) default to 2 teams in lobby correctly read wind value as double

+++ src/de/hhu/propra/team61/MapWindow.java
@@ -157,7 +157,7 @@ public class MapWindow extends Application implements Networkable {
 
         turnCount = input.getInt("turnCount");
         currentTeam = input.getInt("currentTeam");
-        terrain.setWind(input.getInt("windForce"));
+        terrain.setWind(input.getDouble("windForce"));
         initialize();
     }
 
@@ -181,7 +181,7 @@ public class MapWindow extends Application implements Networkable {
 
         turnCount = input.getInt("turnCount");
         currentTeam = input.getInt("currentTeam");
-        terrain.setWind(input.getInt("windForce"));
+        terrain.setWind(input.getDouble("windForce"));
 
         initialize();
